Fix `todo()`; Move `Ligero` implementation to data_structures; minor clippy lints (#31)
* replace direct construction by `new` call to LigeroPCUniversalParams

* some clippy fixes

* replace `todo()`s by `default()` or `()`

* move implementation of `Ligero` to data_structures

* Add more info to `Ligero` struct incl. link and no-hiding note
