+++ src/ligero/data_structures.rs
@@ -1,4 +1,10 @@
+use crate::{
+    Error, PCCommitment, PCCommitterKey, PCPreparedCommitment, PCPreparedVerifierKey, PCRandomness,
+    PCUniversalParams, PCVerifierKey,
+};
 use ark_crypto_primitives::crh::{CRHScheme, TwoToOneCRHScheme};
+use ark_crypto_primitives::merkle_tree::MerkleTree;
+use ark_crypto_primitives::sponge::Absorb;
 use ark_crypto_primitives::{
     merkle_tree::{Config, LeafParam, Path, TwoToOneParam},
     sponge::CryptographicSponge,
@@ -9,16 +15,22 @@ use ark_serialize::{CanonicalDeserialize, CanonicalSerialize};
 use ark_std::fmt::Debug;
 use core::marker::PhantomData;
 use digest::Digest;
-
-use crate::{
-    PCCommitment, PCCommitterKey, PCPreparedCommitment, PCPreparedVerifierKey, PCRandomness,
-    PCUniversalParams, PCVerifierKey,
-};
+use jf_primitives::pcs::transcript::IOPTranscript;
+use std::borrow::Borrow;
 
 use ark_std::rand::RngCore;
 
-// TODO: Disclaimer: no hiding prop
-/// The Ligero polynomial commitment scheme.
+use super::utils::Matrix;
+use super::utils::{
+    calculate_t, compute_dimensions, get_indices_from_transcript, hash_column, reed_solomon,
+};
+
+/// The univariate Ligero polynomial commitment scheme based on [[Ligero]][ligero].
+/// The scheme defaults to the naive batching strategy.
+///
+/// Note: The scheme currently does not support hiding.
+///
+/// [ligero]: https://eprint.iacr.org/2022/1608.pdf
 pub struct Ligero<
     F: PrimeField,
     C: Config,
@@ -31,6 +43,114 @@ pub struct Ligero<
     pub(crate) _digest: PhantomData<D>,
     pub(crate) _sponge: PhantomData<S>,
     pub(crate) _poly: PhantomData<P>,
+}
+
+impl<F, C, D, S, P> Ligero<F, C, D, S, P>
+where
+    F: PrimeField,
+    C: Config,
+    Vec<u8>: Borrow<C::Leaf>,
+    C::InnerDigest: Absorb,
+    D: Digest,
+    S: CryptographicSponge,
+    P: DenseUVPolynomial<F>,
+{
+    /// Create a new instance of Ligero.
+    pub fn new() -> Self {
+        Self {
+            _config: PhantomData,
+            _field: PhantomData,
+            // TODO potentially can get rid of digest and sponge
+            _digest: PhantomData,
+            _sponge: PhantomData,
+            _poly: PhantomData,
+        }
+    }
+
+    pub(crate) fn compute_matrices(polynomial: &P, rho_inv: usize) -> (Matrix<F>, Matrix<F>) {
+        let mut coeffs = polynomial.coeffs().to_vec();
+
+        // 1. Computing parameters and initial matrix
+        let (n_rows, n_cols) = compute_dimensions::<F>(polynomial.degree() + 1); // for 6 coefficients, this is returning 4 x 2 with a row of 0s: fix
+
+        // padding the coefficient vector with zeroes
+        // TODO is this the most efficient/safest way to do it?
+        coeffs.resize(n_rows * n_cols, F::zero());
+
+        let mat = Matrix::new_from_flat(n_rows, n_cols, &coeffs);
+
+        // 2. Apply Reed-Solomon encoding row-wise
+        let ext_mat = Matrix::new_from_rows(
+            mat.rows()
+                .iter()
+                .map(|r| reed_solomon(r, rho_inv))
+                .collect(),
+        );
+
+        (mat, ext_mat)
+    }
+    pub(crate) fn create_merkle_tree(
+        ext_mat: &Matrix<F>,
+        leaf_hash_params: &<<C as Config>::LeafHash as CRHScheme>::Parameters,
+        two_to_one_params: &<<C as Config>::TwoToOneHash as TwoToOneCRHScheme>::Parameters,
+    ) -> MerkleTree<C> {
+        let mut col_hashes: Vec<Vec<u8>> = Vec::new();
+        let ext_mat_cols = ext_mat.cols();
+
+        for col in ext_mat_cols.iter() {
+            col_hashes.push(hash_column::<D, F>(col));
+        }
+
+        // pad the column hashes with zeroes
+        let next_pow_of_two = col_hashes.len().next_power_of_two();
+        col_hashes.resize(next_pow_of_two, vec![0; <D as Digest>::output_size()]);
+
+        MerkleTree::<C>::new(leaf_hash_params, two_to_one_params, col_hashes).unwrap()
+    }
+
+    pub(crate) fn generate_proof(
+        sec_param: usize,
+        rho_inv: usize,
+        b: &[F],
+        mat: &Matrix<F>,
+        ext_mat: &Matrix<F>,
+        col_tree: &MerkleTree<C>,
+        transcript: &mut IOPTranscript<F>,
+    ) -> Result<LigeroPCProofSingle<F, C>, Error> {
+        let t = calculate_t::<F>(sec_param, rho_inv, ext_mat.n)?;
+
+        // 1. left-multiply the matrix by `b`, where for a requested query point `z`,
+        // `b = [1, z^m, z^(2m), ..., z^((m-1)m)]`
+        let v = mat.row_mul(b);
+
+        transcript
+            .append_serializable_element(b"v", &v)
+            .map_err(|_| Error::TranscriptError)?;
+
+        // 2. Generate t column indices to test the linear combination on
+        let indices = get_indices_from_transcript(ext_mat.m, t, transcript)?;
+
+        // 3. Compute Merkle tree paths for the requested columns
+        let mut queried_columns = Vec::with_capacity(t);
+        let mut paths = Vec::with_capacity(t);
+
+        let ext_mat_cols = ext_mat.cols();
+
+        for i in indices {
+            queried_columns.push(ext_mat_cols[i].clone());
+            paths.push(
+                col_tree
+                    .generate_proof(i)
+                    .map_err(|_| Error::TranscriptError)?,
+            );
+        }
+
+        Ok(LigeroPCProofSingle {
+            paths,
+            v,
+            columns: queried_columns,
+        })
+    }
 }
 
 /// The public parameters for the Ligero polynomial commitment scheme.
@@ -223,9 +343,7 @@ where
 pub(crate) type LigeroPCPreparedVerifierKey = ();
 
 impl<Unprepared: PCVerifierKey> PCPreparedVerifierKey<Unprepared> for LigeroPCPreparedVerifierKey {
-    fn prepare(_vk: &Unprepared) -> Self {
-        todo!()
-    }
+    fn prepare(_vk: &Unprepared) -> Self {}
 }
 #[derive(Derivative, CanonicalSerialize, CanonicalDeserialize)]
 #[derivative(Default(bound = ""), Clone(bound = ""), Debug(bound = ""))]
@@ -247,28 +365,24 @@ pub struct LigeroPCCommitment<C: Config> {
 
 impl<C: Config> PCCommitment for LigeroPCCommitment<C> {
     fn empty() -> Self {
-        todo!()
+        LigeroPCCommitment::default()
     }
 
     fn has_degree_bound(&self) -> bool {
-        todo!()
+        false
     }
 }
 
 pub(crate) type LigeroPCPreparedCommitment = ();
 
 impl<Unprepared: PCCommitment> PCPreparedCommitment<Unprepared> for LigeroPCPreparedCommitment {
-    fn prepare(_cm: &Unprepared) -> Self {
-        todo!()
-    }
+    fn prepare(_cm: &Unprepared) -> Self {}
 }
 
 pub(crate) type LigeroPCRandomness = ();
 
 impl PCRandomness for LigeroPCRandomness {
-    fn empty() -> Self {
-        ()
-    }
+    fn empty() -> Self {}
 
     fn rand<R: RngCore>(
         _num_queries: usize,
@@ -276,7 +390,6 @@ impl PCRandomness for LigeroPCRandomness {
         _num_vars: Option<usize>,
         _rng: &mut R,
     ) -> Self {
-        todo!()
     }
 }
